+++ app/javascript/pages/Customers/Show.tsx
@@ -1,0 +1,12 @@
+import { usePage } from "@inertiajs/react";
+import React from "react";
+
+import { default as CustomerDetailPage, CustomerDetailPageProps } from "$app/components/Audience/CustomerDetailPage";
+
+function show() {
+  const props = usePage<CustomerDetailPageProps>().props;
+
+  return <CustomerDetailPage {...props} />;
+}
+
+export default show;
